+++ android/src/main/java/com/bitmovin/player/reactnative/ui/RNPictureInPictureHandler.kt
@@ -7,14 +7,22 @@ import android.util.Log
 import android.util.Rational
 import androidx.annotation.RequiresApi
 import com.bitmovin.player.api.Player
+import com.bitmovin.player.reactnative.PictureInPictureAction
 import com.bitmovin.player.ui.DefaultPictureInPictureHandler
 
 private const val TAG = "RNPiPHandler"
 
+@RequiresApi(Build.VERSION_CODES.O)
 class RNPictureInPictureHandler(
     private val activity: Activity,
     private val player: Player,
 ) : DefaultPictureInPictureHandler(activity, player) {
+    private val pictureInPictureActionHandler = DefaultPictureInPictureActionHandler(
+        activity,
+        player,
+        ::updatePictureInPictureActions,
+    )
+
     // Current PiP implementation on the native side requires playerView.exitPictureInPicture() to be called
     // for `PictureInPictureExit` event to be emitted.
     // Additionally, the event is only emitted if `isPictureInPicture` is true. At the point in time we call
@@ -25,7 +33,10 @@ class RNPictureInPictureHandler(
     override val isPictureInPicture: Boolean
         get() = _isPictureInPicture
 
-    @RequiresApi(Build.VERSION_CODES.O)
+    fun updateActions(actions: List<PictureInPictureAction>) {
+        pictureInPictureActionHandler.updatePictureInPictureActions(actions)
+    }
+
     override fun enterPictureInPicture() {
         if (!isPictureInPictureAvailable) {
             Log.w(TAG, "Calling enterPictureInPicture without PiP support.")
@@ -36,24 +47,33 @@ class RNPictureInPictureHandler(
             return
         }
 
-        // The default implementation doesn't properly handle the case where source isn't loaded yet.
-        // To work around it we just use a 16:9 aspect ratio if we cannot calculate it from `playbackVideoData`.
-        val aspectRatio =
-            player.playbackVideoData
-                ?.let { Rational(it.width, it.height) }
-                ?: Rational(16, 9)
-
-        val params =
-            PictureInPictureParams.Builder()
-                .setAspectRatio(aspectRatio)
-                .build()
-
-        activity.enterPictureInPictureMode(params)
+        activity.enterPictureInPictureMode(buildPictureInPictureParams())
         _isPictureInPicture = true
     }
 
     override fun exitPictureInPicture() {
         super.exitPictureInPicture()
         _isPictureInPicture = false
+    }
+
+    private fun buildPictureInPictureParams(): PictureInPictureParams {
+        // The default implementation doesn't properly handle the case where source isn't loaded yet.
+        // To work around it we just use a 16:9 aspect ratio if we cannot calculate it from `playbackVideoData`.
+        val aspectRatio = player.playbackVideoData
+            ?.let { Rational(it.width, it.height) }
+            ?: Rational(16, 9)
+
+        return PictureInPictureParams.Builder()
+            .setAspectRatio(aspectRatio)
+            .setActions(pictureInPictureActionHandler.buildRemoteActions())
+            .build()
+    }
+
+    private fun updatePictureInPictureActions() {
+        activity.setPictureInPictureParams(buildPictureInPictureParams())
+    }
+
+    fun dispose() {
+        pictureInPictureActionHandler.dispose()
     }
 }