+++ ios/RCTConvert+BitmovinPlayer.swift
@@ -1030,6 +1030,14 @@ extension RCTConvert {
         return pictureInPictureConfig
     }
 
+    static func pictureInPictureActions(_ json: Any?) -> [RNPictureInPictureAction]? {
+        guard let array = json as? [String] else {
+            return nil
+        }
+
+        return array.compactMap(RNPictureInPictureAction.init(rawValue:))
+    }
+
     static func rnPlayerViewConfig(_ json: Any?) -> RNPlayerViewConfig? {
         guard let json = json as? [String: Any?] else {
             return nil
@@ -1183,4 +1191,9 @@ internal struct RNUiConfig {
 internal struct RNBufferLevels {
     let audio: BufferLevel
     let video: BufferLevel
+}
+
+internal enum RNPictureInPictureAction: String {
+    case togglePlayback = "TogglePlayback"
+    case seek = "Seek"
 }