+++ ios/RNPlayerView.swift
@@ -220,6 +220,17 @@ public class RNPlayerView: ExpoView {
         }
         playerView.scalingMode = nativeScalingMode
     }
+
+    internal func updatePictureInPictureActions(actions: [RNPictureInPictureAction]?) {
+        guard let actions else {
+            return
+        }
+        guard let playerView else { return }
+#if os(iOS)
+        playerView.pictureInPicture.showSkipControls = actions.contains(.seek)
+        playerView.tweaks.hidePlaybackControlsInPictureInPicture = !actions.contains(.togglePlayback)
+#endif
+    }
 }
 
 private extension RNPlayerView {