+++ src/components/PlayerView/properties.ts
@@ -4,13 +4,15 @@ import { FullscreenHandler, CustomMessageHandler } from '../../ui';
 import { ScalingMode } from '../../styleConfig';
 import { ViewStyle } from 'react-native';
 import { PlayerViewConfig } from './playerViewConfig';
+import { PictureInPictureAction } from './pictureInPictureAction';
+import { RefObject } from 'react';
 
 /**
  * Base `PlayerView` component props.
  * Used to establish common props between `NativePlayerView` and {@link PlayerView}.
  */
 export interface BasePlayerViewProps {
-  ref?: React.RefObject<null>;
+  ref?: RefObject<null>;
   /**
    * The {@link FullscreenHandler} that is used by the {@link PlayerView} to control the fullscreen mode.
    */
@@ -53,13 +55,31 @@ export interface BasePlayerViewProps {
    * The value must not be altered after setting it initially.
    */
   config?: PlayerViewConfig;
+
+  /**
+   * Picture in Picture actions that should be displayed on the Picture in Picture window.
+   *
+   * Limitations:
+   * - On Android if an empty list is passed and {@link MediaControlConfig.isEnabled} is set to true
+   * play, pause, next, and previous controls will appear, due to the default Android Picture in Picture implementation:
+   * https://developer.android.com/develop/ui/views/picture-in-picture#add_controls
+   * Set {@link MediaControlConfig.isEnabled} to false if this is
+   * not the desired behaviour.
+   * - on iOS/tvOS if {@link PictureInPictureAction.TogglePlayback} is not specified also other actions are
+   * disabled due to OS limitations.
+   *
+   * Default value is `undefined`, which translates to:
+   * - Android: No actions, unless {@link MediaControlConfig.isEnabled} is set to true
+   * - iOS: All actions enabled
+   */
+  pictureInPictureActions?: PictureInPictureAction[];
 }
 
 /**
  * {@link PlayerView} component props.
  */
 export interface PlayerViewProps extends BasePlayerViewProps, PlayerViewEvents {
-  viewRef?: React.MutableRefObject<null>;
+  viewRef?: RefObject<any>;
   /**
    * {@link Player} instance (generally returned from {@link usePlayer} hook) that will control
    * and render audio/video inside the {@link PlayerView}.